Refactor code into separate modules
The main script has been refactored into separate modules 'ai_funcs', 'git_funcs' and 'utils' for readability and maintainability. Code review generation has been separated from git operation functions, and utility functions are moved to a new 'utils' module.
