+++ src/main.rs
@@ -1,175 +1,15 @@
-use async_openai::types::{
-    ChatCompletionRequestAssistantMessageArgs, ChatCompletionRequestSystemMessageArgs,
-    ChatCompletionRequestUserMessageArgs, CreateChatCompletionRequestArgs,
-};
-use clap::{arg, command, Parser};
+mod ai_funcs;
+mod git_funcs;
+mod utils;
+
+use clap::Parser;
 use colored::*;
 use futures::StreamExt;
-use git2::{DiffFormat, DiffOptions, Repository};
 use reqwest;
 use serde::Deserialize;
-use std::io::{stdout, Write};
+use std::io::Write;
 use std::process;
-
-#[derive(Debug, Parser)]
-#[command(author, version, long_about = "AI Code Review Tool")]
-pub struct Args {
-    pub owner: Option<String>,
-    pub repo: Option<String>,
-    pub pr: Option<u32>,
-    #[arg(short, long, action = clap::ArgAction::SetTrue)]
-    pub local: bool,
-}
-
-const COMPLETION_TOKENS: u16 = 1024;
-
-#[derive(Deserialize, Debug)]
-struct File {
-    filename: String,
-    patch: String,
-    status: String,
-}
-
-struct PRInfo {
-    files: Vec<File>,
-}
-
-fn get_git_diff_patch() -> Result<String, git2::Error> {
-    let repo = Repository::open(".")?;
-
-    let mut opts = git2::StatusOptions::new();
-    opts.include_ignored(false);
-
-    let statuses = repo.statuses(Some(&mut opts))?;
-    let mut warned = false;
-    for status in statuses
-        .iter()
-        .filter(|s| s.status() != git2::Status::CURRENT)
-    {
-        let message = format!(
-            "Warning: uncommitted changes detected in file: {}",
-            status.path().unwrap_or("")
-        );
-        println!("{}", message.red());
-        warned = true;
-    }
-
-    if warned {
-        let message = "These changes will be ignored.\n";
-        println!("{}", message.red());
-    }
-
-    let main = repo.revparse_single("main")?;
-    let main_commit = main.peel_to_commit()?;
-    let main_tree = main_commit.tree()?;
-
-    let head = repo.head()?.peel_to_commit()?;
-    let head_tree = head.tree()?;
-
-    let mut opts = DiffOptions::new();
-    let diff = repo.diff_tree_to_tree(Some(&main_tree), Some(&head_tree), Some(&mut opts))?;
-
-    let mut diff_str = String::new();
-    diff.print(DiffFormat::Patch, |_delta, _hunk, line| {
-        let prefix = match line.origin() {
-            '+' => "+",
-            '-' => "-",
-            ' ' => " ",
-            _ => "",
-        };
-        diff_str.push_str(prefix);
-        let content = std::str::from_utf8(line.content()).unwrap_or("");
-        diff_str.push_str(content);
-        true
-    })?;
-    Ok(diff_str)
-}
-
-async fn code_review(output: String) -> Result<(), Box<dyn std::error::Error>> {
-    let client = async_openai::Client::new();
-    let request = CreateChatCompletionRequestArgs::default()
-        .max_tokens(COMPLETION_TOKENS)
-        .model("gpt-4-1106-preview")
-        .messages([
-            ChatCompletionRequestSystemMessageArgs::default()
-                .content("You are a code reviewer. You provide your response in markdown, using a heading (`## path/filename.ext`) for each file reviewed; normal text for your comment; and, potentially, code blocks for code snippets relating to suggested changes (```language...```). Don't bother commenting on everything, just focus on things you think would benefit from being reworked. Very occasionally, you might add positive comments about things that are particularly well executed, but this is entirely optional.")
-                .build()?
-                .into(),
-            ChatCompletionRequestUserMessageArgs::default()
-                .content("Can you review some changes from a PR?")
-                .build()?
-                .into(),
-            ChatCompletionRequestAssistantMessageArgs::default()
-                .content("Sure thing! What are the changes?")
-                .build()?
-                .into(),
-            ChatCompletionRequestUserMessageArgs::default()
-                .content(output.as_str())
-                .build()?
-                .into(),
-        ])
-        .build()?;
-
-    let mut stream = client.chat().create_stream(request).await?;
-
-    println!("== START OF COMMENTS ==\n");
-    let mut lock = stdout().lock();
-    while let Some(result) = stream.next().await {
-        match result {
-            Ok(response) => {
-                response.choices.iter().for_each(|chat_choice| {
-                    if let Some(ref content) = chat_choice.delta.content {
-                        if let Err(e) = write!(lock, "{}", content) {
-                            eprintln!("ERROR WRITING TO STDOUT: {}", e);
-                            process::exit(1)
-                        }
-                    }
-                });
-            }
-            Err(err) => {
-                if let Err(e) = writeln!(lock, "error: {err}") {
-                    return Err(e.into());
-                }
-            }
-        }
-        stdout().flush()?;
-    }
-
-    Ok(())
-}
-
-async fn get_pr_info(
-    owner: &str,
-    repo: &str,
-    pr_number: u32,
-) -> Result<PRInfo, Box<dyn std::error::Error>> {
-    let client = reqwest::Client::new();
-    let pr_url = format!(
-        "https://api.github.com/repos/{}/{}/pulls/{}/files",
-        owner, repo, pr_number
-    );
-
-    let files_info: Vec<File> = client
-        .get(&pr_url)
-        .header("User-Agent", "request")
-        .send()
-        .await?
-        .json()
-        .await?;
-
-    let mut files = Vec::new();
-
-    for file_info in files_info {
-        files.push(file_info);
-    }
-
-    Ok(PRInfo { files })
-}
-
-fn append_with_newline(new_str: &str, buffer: &mut String) {
-    buffer.push_str("\n");
-    buffer.push_str(new_str);
-}
+use utils::Args;
 
 #[tokio::main]
 async fn main() -> Result<(), Box<dyn std::error::Error>> {
@@ -177,9 +17,9 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
 
     if args.local {
         println!("Analysing local changes...\n");
-        let diff = get_git_diff_patch()?;
+        let diff = git_funcs::get_git_diff_patch()?;
         let prompt = format!("\n{}", diff);
-        let review_comments = code_review(prompt).await;
+        let review_comments = ai_funcs::code_review(prompt).await;
         match review_comments {
             Ok(_) => {
                 println!("\n\n== END OF COMMENTS ==");
@@ -208,18 +48,18 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
     });
 
     println!("Getting PR data: {}/{} #{}", owner, repo, pr_number);
-    match get_pr_info(&owner, &repo, pr_number).await {
+    match git_funcs::get_pr_info(&owner, &repo, pr_number).await {
         Ok(pr_info) => {
             let mut output = String::new();
             for file in pr_info.files {
-                append_with_newline(
+                utils::append_with_newline(
                     &format!("{} -- {}", &file.filename, &file.status),
                     &mut output,
                 );
-                append_with_newline(&file.patch, &mut output);
+                utils::append_with_newline(&file.patch, &mut output);
             }
             println!("Analysing changes...\n");
-            let review_result = code_review(output).await;
+            let review_result = ai_funcs::code_review(output).await;
             match review_result {
                 Ok(_) => {
                     println!("\n\n== END OF COMMENTS ==");
